Use `perldocker/perl-tester:latest` image to build our tarball.
This image already has DZil installed in it, so it can help cut down on
the time needed to spin up the initial build and get the tarball made.

After that, though, use a plain/naked Perl image to do the rest of the
testing, to help ensure that we haven't missed adding a dependency
somewhere along the way.